Make Denise per-CCK cycle-accurate (M11.1 fix)
The earlier M11.1 implementation rendered an entire scanline in one
shot at end-of-line. That violates the foundational accuracy rule
(RULES.md 20–21): bitplane DMA arbitration, shift-register reload
timing, and per-pixel modulus updates only emerge from a per-CCK
pipeline, and retrofitting them later is exactly what the rule
forbids.

Denise now ticks once per CCK:
- one bitplane fetch per scheduled lores slot (BPU-driven)
- shift-register reload at slot 7 of each 8-CCK block
- 2 lores pixels output per CCK with per-pixel shift
- end-of-line modulo at hpos==0

The cycle-accurate model has a fetch warm-up: for BPU=1 the first
fetched word does not reach the framebuffer until CCK DDF_START+7.
Update m11_1_bitplanes to sample after that warm-up.

Also fix two pre-existing clippy `unnecessary_cast` errors in
agnus::vhposr that surfaced during the verification pass.

+++ crates/machine-commodore-amiga-ocs/src/agnus.rs
@@ -53,8 +53,8 @@ impl Agnus {
     /// (bits 0-7).
     #[must_use]
     pub fn vhposr(&self) -> u16 {
-        let vpos_lo = (self.vpos & 0xFF) as u16;
-        let hpos = (self.hpos & 0xFF) as u16;
+        let vpos_lo = self.vpos & 0xFF;
+        let hpos = self.hpos & 0xFF;
         (vpos_lo << 8) | hpos
     }
